Collection POST and Individual DELETE implemented
POSTing only makes sense for collections.  It first checks that all the needed
attributes were went along in the request, if they weren't then it responds
with a 400 and a message indicating which attricute was missing.  Otherwise it
goes through and properly creates the entry for the table.  This is done
through the attributes neededNewAttribut and foreign, which then specify what
to look for and how to get the data for the entry.  It then returns the ID
number of the newly created entry in the message.

DELETE is only implemented at the individual level, if needed for the
collection then a series for messages could be sent instead.  It just checks
for the given ID and deletes that instance.